feat: move transactions to background
by moving all transactions logic to the background process, we're enabling multiple sources for transaction details and status.
In this implementation we track all transactions which are made through the extension and update the state from onchain as long as it's not final (reached ACCEPTED_ON_L2)
In case all cases including and after ACCEPTED_ON_L2 it should also show in the voyager history api, and once that's the case we update the details.
This allows for much better transaction tracking, history overview, transaction naming and much more.